requirements.txt
rembg
pillow
streamlit
numpy